Remove FastObjectLoader Dependency on Sequencer
In FastObjectLoader fetch the log globalTail by polling the Log Unit servers instead of relying on
the sequencers view, which in turn needs to wait for the sequencer to
recover in order to provide this. This will allow clients and servers
restarting at the same time to run in parallel.